requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
avro-python3==1.9.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
dill==0.3.1.1
future==0.18.1
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.25.0rc1
h5py==2.10.0
idna==2.8
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
pkg-resources==0.0.0
promise==2.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-datasets==1.3.0
tensorflow-estimator==2.0.1
tensorflow-io-nightly==0.9.0.dev2066
tensorflow-metadata==0.15.0
termcolor==1.1.0
tqdm==4.36.1
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
